Refactor slacklistener to slack subpackage
This will make it easier to expand the Slack functionality over time.

All the message deserialization and routing happens in the routing
module.

Each module in the handlers/ subpackage now implements an individual
handler for a Slack event.
