+++ templatebot/slack/__init__.py
@@ -1,0 +1,6 @@
+"""Listen for Slack events from Kafka.
+"""
+
+__all__ = ('consume_kafka',)
+
+from .router import consume_kafka
