+++ templatebot/slack/handlers/__init__.py
@@ -1,0 +1,4 @@
+from .filedialogsubmission import *
+from .filelisting import *
+from .fileselect import *
+from .projectlisting import *
